feat: add authentitation at the delete-post route, create an unauthorized-error

+++ src/use-cases/delete-post-use-case.ts
@@ -1,6 +1,7 @@
 import { PostsRepository } from "@/repositories/posts-repository"
 import { Post } from "@prisma/client"
 import { ResourceNotFound } from "./errors/resource-not-found-error"
+import { Unauthorized } from "./errors/unauthorized-error"
 
 interface DeletePostUseCaseRequest {
     id: string
@@ -15,12 +16,19 @@ export class DeletePostUseCase {
 
     constructor(private postsRepository: PostsRepository) {}
 
-    async execute({ id }:DeletePostUseCaseRequest): Promise<DeletePostUseCaseResponse> {
-       const post = await this.postsRepository.delete(id)
-        
+    async execute({ id, userId }:DeletePostUseCaseRequest): Promise<DeletePostUseCaseResponse> {
+       const post = await this.postsRepository.findById(id)
+
        if (!post) {
             throw new ResourceNotFound
        }
+
+       if (post.userId !== userId) {
+            throw new Unauthorized()
+       }
+
+       await this.postsRepository.delete(id)
+       
        return { post }
     }
 }
